added eslint and prettier config
