--- src/hackathonWebsites/MelbourneHack/2023/index.css
@@ -1,17 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Inter:wght@100;400;700&display=swap');
-/* 
-* Light font-weight: 100
-* Regular font-weight: 400
-* Bold font-weight: 700
-*/
-
-#melbournehack2023 {
-  color: #ffffff;
-  font-family: 'Inter', sans-serif;
-  background-color: black;
-}
-
-body {
-  background: #000;
-}
-
