+++ tsconfig.node.json
@@ -6,4 +6,4 @@
       "allowSyntheticDefaultImports": true
     },
     "include": ["vite.config.ts"]
-  }
+}
